requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
jinja2==3.1.6
python-dotenv==1.0.1
pycdlib==1.14.0
Pillow==11.1.0
